[FIX] Split large markdown helper file
Split the monolithic src/tools/helpers/markdown.ts (1200+ lines) into
smaller, more maintainable modules:
- markdown.types.ts: Type definitions
- markdown.utils.ts: General helpers, constants, and block creators
- markdown.parser.ts: Markdown to Blocks conversion
- markdown.serializer.ts: Blocks to Markdown conversion

The original src/tools/helpers/markdown.ts now acts as a barrel file
re-exporting everything for backward compatibility.

Tests were also split into:
- src/tools/helpers/markdown.parser.test.ts
- src/tools/helpers/markdown.serializer.test.ts

Verified with 183 tests passing and full type checking.
